backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy[asyncio]==2.0.36
psycopg[binary]==3.2.10
alembic==1.13.2
redis==5.0.8
aiosmtplib==2.0.2
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
Jinja2==3.1.4
Pillow==10.4.0

